requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
bioblend==1.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
dnspython==2.8.0
exceptiongroup==1.3.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
propcache==0.4.1
Pygments==2.19.2
pymongo==4.16.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
tinydb==4.8.2
tuspy==1.1.0
typing_extensions==4.15.0
upstash-vector==0.8.0
urllib3==2.6.3
yarl==1.22.0
langchain==0.3.7
langchain-openai==0.2.7
openai==1.55.3
